requirements.txt
numpy
scipy
pydantic
fastapi
uvicorn[standard]
python-dotenv
pytest
roboticstoolbox-python
openai
spatialmath-python
